New UI for TodoList (#428)
* changed ui for todo lists

* fix error with empty children descedant
