+++ packages/plugins/lists/src/elements/TodoList.tsx
@@ -1,5 +1,7 @@
 import { Elements, PluginElementRenderProps, useBlockData, useYooptaEditor } from '@yoopta/editor';
+import { Check } from 'lucide-react';
 import { TodoListElementProps } from '../types';
+import { cn } from '../utils/cn';
 
 const TodoListRender = ({ extendRender, ...props }: PluginElementRenderProps) => {
   const { attributes, element, children, blockId, HTMLAttributes = {} } = props;
@@ -9,12 +11,7 @@ const TodoListRender = ({ extendRender, ...props }: PluginElementRenderProps) =>
   const { className = '', ...htmlAttrs } = HTMLAttributes;
   const { checked = false } = (element.props || {}) as TodoListElementProps;
 
-  const style = {
-    textDecoration: checked ? 'line-through' : 'none',
-  };
-
   const currentAlign = block?.meta?.align || 'left';
-  const alignClass = `yoopta-align-${currentAlign}`;
 
   if (extendRender) {
     return extendRender(props);
@@ -22,24 +19,46 @@ const TodoListRender = ({ extendRender, ...props }: PluginElementRenderProps) =>
 
   return (
     <div
-      className={`yoopta-todo-list ${alignClass} ${className}`}
+      className={cn('yoopta-todo-list yoo-lists-group', `yoo-lists-text-${currentAlign}`, className)}
       data-checked={checked}
       {...htmlAttrs}
       {...attributes}
     >
-      <span
-        onClick={() => Elements.updateElement(editor, blockId, { type: 'todo-list', props: { checked: !checked } })}
+      <button
+        type="button"
+        onClick={() =>
+          Elements.updateElement(editor, blockId, {
+            type: 'todo-list',
+            props: { checked: !checked },
+          })
+        }
+        className={cn(
+          'yoopta-todo-list-checkbox',
+          checked ? 'yoopta-todo-list-checkbox--checked' : 'yoopta-todo-list-checkbox--unchecked',
+        )}
         contentEditable={false}
-        className="yoopta-todo-list-checkbox"
       >
+        <Check
+          className={cn(
+            'yoopta-todo-list-checkbox-icon',
+            checked ? 'yoopta-todo-list-checkbox-icon--checked' : 'yoopta-todo-list-checkbox-icon--unchecked',
+          )}
+        />
         <input
           type="checkbox"
           className="yoopta-todo-list-checkbox-input"
           checked={checked}
           disabled={editor.readOnly}
+          onChange={() => {}}
         />
-      </span>
-      <div className="yoopta-todo-list-content" style={style}>
+      </button>
+
+      <div
+        className={cn('yoopta-todo-list-content', checked && 'yoopta-todo-list-content--checked')}
+        style={{
+          textDecoration: checked ? 'line-through' : 'none',
+        }}
+      >
         {children}
       </div>
     </div>
