+++ packages/plugins/lists/src/plugin/NumberedList.tsx
@@ -1,16 +1,9 @@
-import {
-  YooptaPlugin,
-  YooptaBlockData,
-  generateId,
-  deserializeTextNodes,
-  serializeTextNodes,
-  serializeTextNodesIntoMarkdown,
-  Blocks, deserializeListNodes,
-} from '@yoopta/editor';
+import { YooptaPlugin, YooptaBlockData, serializeTextNodes, serializeTextNodesIntoMarkdown } from '@yoopta/editor';
 import { NumberedListCommands } from '../commands';
 import { NumberedListRender } from '../elements/NumberedList';
 import { onKeyDown } from '../events/onKeyDown';
 import { ListElementMap } from '../types';
+import { deserializeListNodes } from '../utils/deserializeListNodes';
 
 const NumberedList = new YooptaPlugin<Pick<ListElementMap, 'numbered-list'>>({
   type: 'NumberedList',
