+++ packages/plugins/lists/src/plugin/TodoList.tsx
@@ -1,16 +1,9 @@
-import {
-  buildBlockData, deserializeListNodes,
-  deserializeTextNodes,
-  generateId,
-  serializeTextNodes,
-  serializeTextNodesIntoMarkdown,
-  YooptaBlockData,
-  YooptaPlugin,
-} from '@yoopta/editor';
+import { serializeTextNodes, serializeTextNodesIntoMarkdown, YooptaBlockData, YooptaPlugin } from '@yoopta/editor';
 import { TodoListCommands } from '../commands';
 import { TodoListRender } from '../elements/TodoList';
 import { onKeyDown } from '../events/onKeyDown';
 import { ListElementMap } from '../types';
+import { deserializeListNodes } from '../utils/deserializeListNodes';
 
 const TodoList = new YooptaPlugin<Pick<ListElementMap, 'todo-list'>>({
   type: 'TodoList',
