+++ packages/plugins/lists/src/utils/cn.ts
@@ -1,0 +1,3 @@
+export function cn(...classes: (string | undefined | boolean | null)[]) {
+  return classes.filter(Boolean).join(' ');
+}
